Update network responses to support 64-bit Job IDs

+++ common/network_test.go
@@ -215,7 +215,7 @@ func TestSecrets_expandVariables(t *testing.T) {
 }
 
 func TestJobResponse_JobURL(t *testing.T) {
-	jobID := 1
+	jobID := int64(1)
 	//nolint:lll
 	testCases := map[string]string{
 		"http://user:pass@gitlab.example.com/my-namespace/my-project.git":     "http://gitlab.example.com/my-namespace/my-project/-/jobs/1",
